Remove load_expert_states
